# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KenyaPay Payroll System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll/runs:
    get:
      operationId: listPayrollRuns
      tags: [payroll]
      summary: List all payroll runs
      responses:
        "200":
          description: List of payroll runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRun"
    post:
      operationId: createPayrollRun
      tags: [payroll]
      summary: Create and process a new payroll run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}:
    get:
      operationId: getPayrollRun
      tags: [payroll]
      summary: Get a payroll run by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}/entries:
    get:
      operationId: listPayrollEntries
      tags: [payroll]
      summary: List all payroll entries for a run
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollEntry"
  /payroll/runs/{id}/approve:
    patch:
      operationId: approvePayrollRun
      tags: [payroll]
      summary: Approve a payroll run
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovePayrollInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}/disburse:
    patch:
      operationId: disbursePayrollRun
      tags: [payroll]
      summary: Disburse salaries for a payroll run
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisbursePayrollInput"
      responses:
        "200":
          description: Disbursed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
...
  /payroll/tax-preview:
    post:
      operationId: previewTaxCalculation
      tags: [payroll]
      summary: Preview tax and deduction calculation for a given gross salary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxPreviewInput"
      responses:
        "200":
          description: Tax preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxPreview"
...
components:
  schemas:
...
    PayrollRun:
      type: object
      required:
        - id
        - month
        - year
        - status
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
          enum: [draft, processing, approved, disbursed]
        totalGross:
          type: number
        totalPaye:
          type: number
        totalNssf:
          type: number
        totalShif:
          type: number
        totalHousingLevy:
          type: number
        totalNet:
          type: number
        employeeCount:
          type: integer
        processedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        disbursedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PayrollRunInput:
      type: object
      required:
        - month
        - year
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
    ApprovePayrollInput:
      type: object
      properties:
        notes:
          type: string
    DisbursePayrollInput:
      type: object
      properties:
        notes:
          type: string
    PayrollEntry:
      type: object
      required:
        - id
        - payrollRunId
        - employeeId
        - grossSalary
        - netPay
      properties:
        id:
          type: integer
        payrollRunId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeNumber:
          type: string
        departmentName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        grossSalary:
          type: number
        basicSalary:
          type: number
        paye:
          type: number
        nssfEmployee:
          type: number
        nssfEmployer:
          type: number
        shif:
          type: number
        housingLevyEmployee:
          type: number
        housingLevyEmployer:
          type: number
        personalRelief:
          type: number
        insuranceRelief:
          type: number
        otherDeductions:
          type: number
        netPay:
          type: number
        paymentMethod:
          type: string
        disbursementStatus:
          type: string
          enum: [pending, processing, success, failed]
        month:
          type: integer
        year:
          type: integer
...
    TaxPreviewInput:
      type: object
      required:
        - grossSalary
      properties:
        grossSalary:
          type: number
    TaxPreview:
      type: object
      properties:
        grossSalary:
          type: number
        nssfEmployee:
          type: number
        nssfEmployer:
          type: number
        shif:
          type: number
        housingLevyEmployee:
          type: number
        housingLevyEmployer:
          type: number
        taxableIncome:
          type: number
        paye:
          type: number
        personalRelief:
          type: number
        insuranceRelief:
          type: number
        netPaye:
          type: number
        netPay:
          type: number
        totalEmployerCost:
          type: number
...